postgres_to_es/requirements.txt
psycopg2-binary==2.9.1
requests==2.26.0
backoff==1.11.1
pydantic==1.8.2
python-dotenv==0.19.0
elasticsearch==7.14.1
redis